+++ src/components/ClearButton/index.tsx
@@ -1,15 +1,13 @@
-const ClearButton = () => {
+import { IfcClearButton } from '../../types';
+
+const ClearButton = (clearButtonProps: IfcClearButton) => {
+  const { handleClear } = clearButtonProps;
   const clearStatus = () => {
-    if (typeof Storage !== 'undefined') {
-      localStorage.removeItem('ElectoralCollegeStatus');
-      localStorage.removeItem('PopularVoteStatus');
-      localStorage.removeItem('PopularVoteTotals');
-      localStorage.removeItem('WinnerTakeAllTotals');
-      alert('Cleared!');
-      location.reload(); // eslint-disable-line no-restricted-globals
-    } else {
-      alert('Please enable Storage to save your selections.');
-    }
+    localStorage.removeItem('ElectoralCollegeStatus');
+    localStorage.removeItem('PopularVoteStatus');
+    localStorage.removeItem('PopularVoteTotals');
+    localStorage.removeItem('WinnerTakeAllTotals');
+    handleClear();
   };
 
   return (